Logo inicial, alerta refeito e hud melhorada
E portas abrindo de acordo com a qtd de inimigos mortos
Hack: F para FasrFoward

+++ Refidenptio/Assets/Scripts/UI/UITextAmmo.cs
@@ -6,8 +6,8 @@ using TMPro;
 
 public class UITextAmmo : MonoBehaviour
 {
-    public int qtdAmmo;
-    public int maxAmmo;
+    public float qtdAmmo;
+    public float maxAmmo;
 
     private TextMeshProUGUI ammoText;
